[console] - cleaned up range toolbar context menu
