+++ console/src/range/external.ts
@@ -11,7 +11,6 @@ import { Layout } from "@/layout";
 import { Edit, EDIT_LAYOUT_TYPE } from "@/range/EditLayout";
 import { MetaData, metaDataWindowLayout } from "@/range/MetaData";
 
-export * from "@/range/accordionEntry";
 export * from "@/range/ContextMenu";
 export * from "@/range/EditLayout";
 export * from "@/range/link";
